refactor: replace dependency gitpkg.vercel.app dependencies
